Complete refactor of New Task component and some file house keeping
